moved the events to a custom class moved the event to BrowserVC instead of AppDelegate
