+++ firefox-ios/Client/Frontend/Browser/BrowserViewController/Views/BrowserViewController.swift
@@ -71,6 +71,7 @@ class BrowserViewController: UIViewController,
     // MARK: Telemetry Variables
     var webviewTelemetry = WebViewLoadMeasurementTelemetry()
     var privateBrowsingTelemetry = PrivateBrowsingTelemetry()
+    var appStartupTelemetry = AppStartupTelemetry()
 
     // popover rotation handling
     var displayedPopoverController: UIViewController?
@@ -2744,6 +2745,7 @@ extension BrowserViewController {
     func trackTelemetry() {
         trackAccessibility()
         trackNotificationPermission()
+        appStartupTelemetry.sendStartupTelemetry()
     }
 
     func trackAccessibility() {
